# Write state using compact JSON representation