--- a/internal/command/apply_test.go
+++ b/internal/command/apply_test.go
@@ -1594,7 +1594,7 @@ func TestApply_backup(t *testing.T) {
 				Name: "foo",
 			}.Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
 			&states.ResourceInstanceObjectSrc{
-				AttrsJSON: []byte("{\n            \"id\": \"bar\"\n          }"),
+				AttrsJSON: []byte("{\"id\":\"bar\"}"),
 				Status:    states.ObjectReady,
 			},
 			addrs.AbsProviderConfig{