--- a/internal/states/statefile/version4.go
+++ b/internal/states/statefile/version4.go
@@ -9,7 +9,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"os"
 	"sort"
 
 	version "github.com/hashicorp/go-version"
@@ -425,13 +424,7 @@ func writeStateV4(file *File, w io.Writer, enc encryption.StateEncryption) tfdia
 
 	sV4.normalize()
 
-	var src []byte
-	var err error
-	if _, exists := os.LookupEnv("TF_USE_COMPACT_STATE_FORMAT"); exists {
-		src, err = json.Marshal(sV4)
-	} else {
-		src, err = json.MarshalIndent(sV4, "", "  ")
-	}
+	src, err := json.Marshal(sV4)
 	if err != nil {
 		// Shouldn't happen if we do our conversion to *stateV4 correctly above.
 		diags = diags.Append(tfdiags.Sourceless(