… internal/command/testdata/state-push-good/replace.tfstate
… internal/command/testdata/state-push-replace-match/local-state.tfstate
… internal/command/testdata/state-push-replace-match/replace.tfstate
… internal/command/testdata/state-push-serial-older/local-state.tfstate
… internal/command/testdata/state-push-serial-older/replace.tfstate
… internal/states/statefile/testdata/roundtrip/v1-simple.in.tfstate
… internal/states/statefile/testdata/roundtrip/v1-simple.out.tfstate
… internal/states/statefile/testdata/roundtrip/v3-bigint.in.tfstate
… internal/states/statefile/testdata/roundtrip/v3-bigint.out.tfstate
… internal/states/statefile/testdata/roundtrip/v3-builtin.in.tfstate
… internal/states/statefile/testdata/roundtrip/v3-builtin.out.tfstate
… internal/states/statefile/testdata/roundtrip/v3-grabbag.in.tfstate
… internal/states/statefile/testdata/roundtrip/v3-grabbag.out.tfstate
… internal/states/statefile/testdata/roundtrip/v3-invalid-depends.in.tfstate
… internal/states/statefile/testdata/roundtrip/v3-invalid-depends.out.tfstate
… internal/states/statefile/testdata/roundtrip/v3-simple.in.tfstate
… internal/states/statefile/testdata/roundtrip/v3-simple.out.tfstate
… internal/states/statefile/testdata/roundtrip/v4-cbd.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-foreach.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-future.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-foreach.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-foreach.out.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-modules.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-modules.out.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-simple.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-legacy-simple.out.tfstate
… internal/states/statefile/testdata/roundtrip/v4-modules.in.tfstate
… internal/states/statefile/testdata/roundtrip/v4-simple.in.tfstate