@cam72cam 2024-05-21
I had a small heart attack when looking at the LoC totals in the diff, before remembering that it was the whole goal of this PR :laughing: