@eduzgun 2024-06-09
@cam72cam Some of the testdata in `/internal/states/statefile/testdata/roundtrip` looks wrong and is causing ` go test ./internal/states/statefile` tests to fail. These files are: v4-cbd.out.tfstate, v4-foreach.out.tfstate, v4-modules.out.tfstate, v4-simple.in.tfstate. Should I create a new issue for this?