@cam72cam 2024-06-09
@eduzgun, sure create an issue.  I can't reproduce that myself, the tests are passing on main.