@bobrik 2024-08-27
This is somewhat annoying for simple use-cases where a file is saved locally. Previously the changes could be reasonable diffed and saved into git, now it's a single gigantic line that cannot be viewed as a diff. It's evident just from the updated testdata itself:

* https://github.com/opentofu/opentofu/commit/522a859cf012ab2f1e2a42355281d2f7ad961c28#diff-e184610507824cd7d0b832b4e3b9267e88df76f63bd2b07e66c33157318ef6a0

Further changes will be even less readable.

I think there's a middle ground here, where a compact representation can be used for non-local backends (or perhaps if there is enough data for it to matter).