@cam72cam 2024-08-27
Could you create an issue linking to this PR @bobrik?  I'm curious what patterns you follow that makes a git diff of the state file useful.